src/ai/pyproject.toml
[project]
name = "lingua"
version = "0.1.0"
description = "an AI powered app for learning languages with AI"
readme = "README.md"
requires-python = ">=3.12.0"
dependencies = [
    "google-adk>=1.26.0",
    "google-cloud-aiplatform>=1.133.0",
    "google-cloud-firestore>=2.0.0",
    "pydantic>=2.12.5",
    "pytz>=2025.2",
    "typer>=0.24.1",
    "vertexai>=1.43.0",
]
